+++ NEWS.md
@@ -1,4 +1,4 @@
-# rsf 0.3.0
+# rsf (development version)
 
 * Use GitHub Actions for CI and pkgdown site
 * Add project options to initialize with git repository and/or renv private library